murdermysterycase-query.sql
OVERVIEW:

On the 15th of January 2018, a crime took place in SQL City and help is needed by the detective to solve this case. A crime scene report was given to me but I somehow lost it.  I vaguely remember certain details like the date and the city as previously mentioned.  To figure out the culprit, I started by getting information and the corresponding report for the police department database.

All analysis of crimes committed on the 15th of January in SQL city from the police department database is highlighted in this report.

 Below are the steps taken to figure out who committed the crime